+++ src/app/(app)/chamados/criar/ticket-create/ticket-create-form.tsx
@@ -1,0 +1,1020 @@
+'use client'
+
+import { ChevronDown, Plus, SquareCheck, Trash, Upload } from 'lucide-react'
+import { Controller } from 'react-hook-form'
+
+import { Button } from '@/components/ui/button'
+import {
+  Command,
+  CommandEmpty,
+  CommandGroup,
+  CommandInput,
+  CommandItem,
+  CommandList,
+} from '@/components/ui/command'
+import { Input } from '@/components/ui/input'
+import { Label } from '@/components/ui/label'
+import {
+  Popover,
+  PopoverContent,
+  PopoverTrigger,
+} from '@/components/ui/popover'
+import {
+  Select,
+  SelectContent,
+  SelectItem,
+  SelectTrigger,
+  SelectValue,
+} from '@/components/ui/select'
+import { Switch } from '@/components/ui/switch'
+import { Textarea } from '@/components/ui/textarea'
+
+import { CorrelataListForm } from '../components/services/correlata-list-form'
+import { ServiceAddCard } from '../components/services/service-add-card'
+import { ServiceList } from '../components/services/service-list'
+import { ServiceModal } from '../components/services/service-modal'
+import { DataBaseDatePicker } from '../components/shared/data-base-date-picker'
+import { PriorityButton } from '../components/shared/priority-button'
+import { Section } from '../components/shared/section'
+import { useTicketCreate } from '../hooks/use-ticket-create'
+import { MOCK_TEAMS } from '../ticket-create/ticket-create.constant'
+import styles from './ticket-create-form.module.css'
+
+function nullIfEmpty(value?: string | null) {
+  if (value == null) return null
+  const trimmed = value.trim()
+  return trimmed.length ? trimmed : null
+}
+
+export function TicketCreateForm() {
+  const vm = useTicketCreate()
+  const initialBuscaPorPlaca =
+    vm.serviceModalOpen === 'busca_por_placa' &&
+    vm.serviceModalEditIndex !== null
+      ? vm.getValues().busca_por_placa?.[vm.serviceModalEditIndex]
+      : undefined
+
+  const initialBuscaPorRadar =
+    vm.serviceModalOpen === 'busca_por_radar' &&
+    vm.serviceModalEditIndex !== null
+      ? vm.getValues().busca_por_radar?.[vm.serviceModalEditIndex]
+      : undefined
+
+  const initialCerco =
+    vm.serviceModalOpen === 'cerco_eletronico' &&
+    vm.serviceModalEditIndex !== null
+      ? vm.getValues().cerco_eletronico?.[vm.serviceModalEditIndex]
+      : undefined
+
+  const initialBuscaPorImagem =
+    vm.serviceModalOpen === 'busca_por_imagem' &&
+    vm.serviceModalEditIndex !== null
+      ? vm.getValues().busca_por_imagem?.[vm.serviceModalEditIndex]
+      : undefined
+
+  const initialPlacasCorrelatas =
+    vm.serviceModalOpen === 'placas_correlatas' &&
+    vm.serviceModalEditIndex !== null
+      ? vm.getValues().placas_correlatas?.[vm.serviceModalEditIndex]
+      : undefined
+
+  const initialPlacasConjuntas =
+    vm.serviceModalOpen === 'placas_conjuntas' &&
+    vm.serviceModalEditIndex !== null
+      ? vm.getValues().placas_conjuntas?.[vm.serviceModalEditIndex]
+      : undefined
+
+  const initialReservaImagem =
+    vm.serviceModalOpen === 'reserva_de_imagem' &&
+    vm.serviceModalEditIndex !== null
+      ? vm.getValues().reserva_de_imagem?.[vm.serviceModalEditIndex]
+      : undefined
+
+  const initialAnaliseImagem =
+    vm.serviceModalOpen === 'analise_de_imagem' &&
+    vm.serviceModalEditIndex !== null
+      ? vm.getValues().analise_de_imagem?.[vm.serviceModalEditIndex]
+      : undefined
+
+  const initialOutros =
+    vm.serviceModalOpen === 'outros' && vm.serviceModalEditIndex !== null
+      ? vm.getValues().outros?.[vm.serviceModalEditIndex]
+      : undefined
+
+  return (
+    <div className={styles.root}>
+      <div className="mx-auto w-full max-w-5xl space-y-8">
+        <h1
+          className="font-semibold leading-10 text-[var(--tc-heading,#f9fafa)]"
+          style={{ fontSize: '20px' }}
+        >
+          Cadastro Manual de Chamado
+        </h1>
+        <p className="mt-0 text-[length:12px] leading-4 text-[var(--tc-muted,#97a2ab)]">
+          Preencha as informações abaixo para criar um novo chamado.
+        </p>
+      </div>
+
+      <form
+        className="mx-auto w-full max-w-5xl space-y-8"
+        onSubmit={vm.handleSubmit(vm.onSubmit)}
+      >
+        <div className={`${styles.sectionCard} ${styles.sectionCardFirst}`}>
+          <div className="grid grid-cols-1 gap-4 md:grid-cols-2">
+            <div className="space-y-1.5">
+              <Label className={styles.fieldLabel}>Associar chamado</Label>
+
+              <Controller
+                control={vm.control}
+                name="associar_chamado_id"
+                render={({ field }) => (
+                  <Popover
+                    open={vm.ticketPopoverOpen}
+                    onOpenChange={vm.setTicketPopoverOpen}
+                  >
+                    <PopoverTrigger asChild>
+                      <Button
+                        type="button"
+                        variant="outline"
+                        disabled={vm.isLoading}
+                        className={`h-11 w-full justify-between ${styles.inputBg} font-normal`}
+                      >
+                        <span className="truncate">
+                          {vm.selectedTicketLabel ||
+                            'Digite para buscar um chamado'}
+                        </span>
+                        <ChevronDown className="h-4 w-4 opacity-50" />
+                      </Button>
+                    </PopoverTrigger>
+
+                    <PopoverContent
+                      className={`w-[var(--radix-popover-trigger-width)] p-0 ${styles.associarChamadoPopover}`}
+                    >
+                      <Command
+                        shouldFilter={false}
+                        className={styles.associarChamadoCommand}
+                      >
+                        <CommandInput
+                          placeholder="Buscar chamado..."
+                          value={vm.ticketSearch}
+                          onValueChange={vm.setTicketSearch}
+                        />
+
+                        <CommandList>
+                          {vm.isTicketsLoading && (
+                            <div className="px-3 py-2 text-sm text-muted-foreground">
+                              Buscando chamados...
+                            </div>
+                          )}
+
+                          {!vm.isTicketsLoading &&
+                            vm.ticketSearch.trim().length === 0 && (
+                              <div className="px-3 py-2 text-sm text-muted-foreground">
+                                Digite para pesquisar.
+                              </div>
+                            )}
+
+                          {!vm.isTicketsLoading &&
+                            vm.ticketSearch.trim().length > 0 &&
+                            vm.tickets.length === 0 && (
+                              <CommandEmpty>
+                                Nenhum chamado encontrado.
+                              </CommandEmpty>
+                            )}
+
+                          <CommandGroup>
+                            {vm.tickets.map((ticket) => (
+                              <CommandItem
+                                key={ticket.id}
+                                value={`${ticket.id} ${ticket.titulo}`}
+                                onSelect={() => {
+                                  field.onChange(ticket.id)
+                                  vm.setSelectedTicketLabel(ticket.titulo)
+                                  vm.setTicketSearch(ticket.titulo)
+                                  vm.setTicketPopoverOpen(false)
+                                }}
+                              >
+                                <div className="flex w-full items-center justify-between gap-2">
+                                  <span className="truncate">
+                                    {ticket.titulo}
+                                  </span>
+                                  {field.value === ticket.id && (
+                                    <span className="text-xs text-muted-foreground">
+                                      Selecionado
+                                    </span>
+                                  )}
+                                </div>
+                              </CommandItem>
+                            ))}
+                          </CommandGroup>
+
+                          {field.value && (
+                            <div className="border-t p-2">
+                              <Button
+                                type="button"
+                                variant="ghost"
+                                className="w-full justify-start"
+                                onClick={() => {
+                                  field.onChange(null)
+                                  vm.setSelectedTicketLabel('')
+                                  vm.setTicketSearch('')
+                                  vm.setTicketPopoverOpen(false)
+                                }}
+                              >
+                                Limpar seleção
+                              </Button>
+                            </div>
+                          )}
+                        </CommandList>
+                      </Command>
+                    </PopoverContent>
+                  </Popover>
+                )}
+              />
+            </div>
+
+            <div className="space-y-1.5">
+              <Label className={styles.fieldLabel}>Tipo de chamado</Label>
+              <Controller
+                control={vm.control}
+                name="tipo_chamado_id"
+                render={({ field }) => (
+                  <Select
+                    value={field.value}
+                    onValueChange={field.onChange}
+                    disabled={vm.isLoading || vm.isTicketTypesLoading}
+                  >
+                    <SelectTrigger className={`h-11 ${styles.inputBg}`}>
+                      <SelectValue placeholder="Selecione" />
+                    </SelectTrigger>
+                    <SelectContent className={styles.selectContentForm}>
+                      {vm.ticketTypes.map((ticketType) => (
+                        <SelectItem
+                          key={ticketType.id}
+                          value={ticketType.id}
+                          className={styles.selectItemForm}
+                        >
+                          {ticketType.name}
+                        </SelectItem>
+                      ))}
+                    </SelectContent>
+                  </Select>
+                )}
+              />
+              {vm.errors.tipo_chamado_id?.message && (
+                <p className="text-xs text-destructive">
+                  {vm.errors.tipo_chamado_id.message}
+                </p>
+              )}
+            </div>
+          </div>
+        </div>
+
+        <Section
+          title="Informações do chamado"
+          isOpen={vm.openSections.info}
+          onToggle={() => vm.toggleSection('info')}
+        >
+          <div className="grid grid-cols-1 gap-3 md:grid-cols-2">
+            <div className="space-y-1.5">
+              <Label className={styles.fieldLabel}>Nº de procedimento</Label>
+              <Input
+                className={`h-11 ${styles.inputBg}`}
+                disabled={vm.isLoading}
+                {...vm.register('numero_procedimento')}
+              />
+            </div>
+
+            <div className="space-y-1.5">
+              <Label className={styles.fieldLabel}>Nº do ofício</Label>
+              <Input
+                className={`h-11 ${styles.inputBg}`}
+                disabled={vm.isLoading}
+                {...vm.register('numero_oficio')}
+              />
+            </div>
+
+            <div className="space-y-1.5">
+              <Label className={styles.fieldLabel}>Data base</Label>
+              <Controller
+                control={vm.control}
+                name="data_base"
+                render={({ field }) => (
+                  <DataBaseDatePicker
+                    value={field.value ?? ''}
+                    onChange={field.onChange}
+                    disabled={vm.isLoading}
+                  />
+                )}
+              />
+            </div>
+
+            <div className="space-y-1.5">
+              <Label className={styles.fieldLabel}>Natureza</Label>
+              <Controller
+                control={vm.control}
+                name="natureza_id"
+                render={({ field }) => (
+                  <Select
+                    value={field.value ?? ''}
+                    onValueChange={(v) => field.onChange(v || null)}
+                    disabled={vm.isLoading || vm.isTicketNaturesLoading}
+                  >
+                    <SelectTrigger className={`h-11 ${styles.inputBg}`}>
+                      <SelectValue placeholder="Selecione" />
+                    </SelectTrigger>
+                    <SelectContent className={styles.selectContentForm}>
+                      {vm.ticketNatures.map((nature) => (
+                        <SelectItem
+                          key={nature.id}
+                          value={nature.id}
+                          className={styles.selectItemForm}
+                        >
+                          {nature.name}
+                        </SelectItem>
+                      ))}
+                    </SelectContent>
+                  </Select>
+                )}
+              />
+            </div>
+
+            <div className="flex flex-wrap items-center gap-2 pt-2 md:col-span-2">
+              <Label className={styles.fieldLabel}>
+                Possui apelido pela imprensa?
+              </Label>
+
+              <Controller
+                control={vm.control}
+                name="possui_apelido_imprensa"
+                render={({ field }) => (
+                  <div className="flex items-center gap-2">
+                    <Switch
+                      id="possui_apelido_imprensa"
+                      size="sm"
+                      checked={!!field.value}
+                      onCheckedChange={field.onChange}
+                      disabled={vm.isLoading}
+                      className={styles.apelidoToggle}
+                    />
+                    <span className={styles.toggleLabel}>
+                      {field.value ? 'Sim' : 'Não'}
+                    </span>
+                  </div>
+                )}
+              />
+            </div>
+
+            {vm.possuiApelido && (
+              <>
+                <div className="space-y-1.5">
+                  <Label className={styles.fieldLabel}>
+                    Apelido pela imprensa
+                  </Label>
+                  <Input
+                    className={`h-11 ${styles.inputBg}`}
+                    disabled={vm.isLoading}
+                    {...vm.register('apelido_imprensa')}
+                  />
+                </div>
+
+                <div className="space-y-1.5">
+                  <Label className={styles.fieldLabel}>Link da matéria</Label>
+                  <Input
+                    className={`h-11 ${styles.inputBg}`}
+                    disabled={vm.isLoading}
+                    placeholder="https://..."
+                    {...vm.register('link_materia')}
+                  />
+                  {vm.errors.link_materia?.message && (
+                    <p className="text-xs text-destructive">
+                      {vm.errors.link_materia.message}
+                    </p>
+                  )}
+                </div>
+              </>
+            )}
+          </div>
+        </Section>
+
+        <Section
+          title="Requisitante"
+          isOpen={vm.openSections.requester}
+          onToggle={() => vm.toggleSection('requester')}
+        >
+          <div className="space-y-3">
+            <div className="space-y-1.5">
+              <Label className={styles.fieldLabel}>Demandante</Label>
+              <Controller
+                control={vm.control}
+                name="operation_id"
+                render={({ field }) => (
+                  <Select
+                    value={field.value ?? ''}
+                    onValueChange={(v) => field.onChange(v || null)}
+                    disabled={vm.isLoading || vm.isOperationsLoading}
+                  >
+                    <SelectTrigger className={`h-11 ${styles.inputBg}`}>
+                      <SelectValue placeholder="Selecione" />
+                    </SelectTrigger>
+                    <SelectContent className={styles.selectContentForm}>
+                      {vm.operations.map((op) => (
+                        <SelectItem
+                          key={op.id}
+                          value={op.id}
+                          className={styles.selectItemForm}
+                        >
+                          {op.title}
+                        </SelectItem>
+                      ))}
+                    </SelectContent>
+                  </Select>
+                )}
+              />
+              {vm.errors.operation_id?.message && (
+                <p className="text-xs text-destructive">
+                  {vm.errors.operation_id.message}
+                </p>
+              )}
+            </div>
+
+            <div className="grid grid-cols-1 gap-3 md:grid-cols-3">
+              <div className="space-y-1.5">
+                <Label className={styles.fieldLabel}>Requisitante</Label>
+                <Input
+                  className={`h-11 ${styles.inputBg}`}
+                  disabled={vm.isLoading}
+                  {...vm.register('requisitante.requisitante_nome')}
+                />
+                {vm.errors.requisitante?.requisitante_nome?.message && (
+                  <p className="text-xs text-destructive">
+                    {vm.errors.requisitante?.requisitante_nome?.message}
+                  </p>
+                )}
+              </div>
+
+              <div className="space-y-1.5">
+                <Label className={styles.fieldLabel}>Telefone</Label>
+                <Input
+                  className={`h-11 ${styles.inputBg}`}
+                  disabled={vm.isLoading}
+                  {...vm.register('requisitante.requisitante_telefone')}
+                />
+              </div>
+
+              <div className="space-y-1.5">
+                <Label className={styles.fieldLabel}>Email</Label>
+                <Input
+                  className={`h-11 ${styles.inputBg}`}
+                  disabled={vm.isLoading}
+                  {...vm.register('requisitante.requisitante_email')}
+                />
+                {vm.errors.requisitante?.requisitante_email?.message && (
+                  <p className="text-xs text-destructive">
+                    {vm.errors.requisitante?.requisitante_email?.message}
+                  </p>
+                )}
+              </div>
+            </div>
+
+            <div className="flex flex-col items-end gap-4">
+              <button
+                type="button"
+                disabled={vm.isLoading || vm.focalPoints.fields.length >= 20}
+                onClick={() =>
+                  vm.focalPoints.append({
+                    nome: '',
+                    telefone: '',
+                    email: null,
+                  })
+                }
+                className={styles.addPointFocalButton}
+              >
+                <Plus className="h-5 w-5 shrink-0" aria-hidden />
+                Adicionar Ponto Focal
+              </button>
+            </div>
+
+            <div className="space-y-3">
+              {vm.focalPoints.fields.map((fp, idx) => (
+                <div
+                  key={fp.id}
+                  className="grid grid-cols-1 gap-3 md:grid-cols-3"
+                >
+                  <div className="space-y-1.5">
+                    <Label className={styles.fieldLabel}>Ponto focal</Label>
+                    <Input
+                      className={`h-11 ${styles.inputBg}`}
+                      disabled={vm.isLoading}
+                      {...vm.register(`pontos_focais.${idx}.nome`)}
+                    />
+                  </div>
+
+                  <div className="space-y-1.5">
+                    <Label className={styles.fieldLabel}>Telefone</Label>
+                    <Input
+                      className={`h-11 ${styles.inputBg}`}
+                      disabled={vm.isLoading}
+                      {...vm.register(`pontos_focais.${idx}.telefone`)}
+                    />
+                  </div>
+
+                  <div className="space-y-1.5">
+                    <Label className={styles.fieldLabel}>Email</Label>
+                    <div className="flex gap-2">
+                      <Input
+                        className={`h-11 ${styles.inputBg}`}
+                        disabled={vm.isLoading}
+                        {...vm.register(`pontos_focais.${idx}.email`)}
+                      />
+                      <Button
+                        type="button"
+                        variant="ghost"
+                        className="h-11 w-11 p-0"
+                        disabled={vm.isLoading}
+                        onClick={() => vm.focalPoints.remove(idx)}
+                      >
+                        <Trash className="h-4 w-4" />
+                      </Button>
+                    </div>
+                  </div>
+                </div>
+              ))}
+            </div>
+          </div>
+        </Section>
+
+        <Section
+          title="Serviços"
+          isOpen={vm.openSections.services}
+          onToggle={() => vm.toggleSection('services')}
+        >
+          <div className={styles.serviceGrid}>
+            <ServiceAddCard
+              title="Busca por placa"
+              onAdd={() => vm.handleOpenService('busca_por_placa')}
+            />
+            <ServiceAddCard
+              title="Busca por radar"
+              onAdd={() => vm.handleOpenService('busca_por_radar')}
+            />
+            <ServiceAddCard
+              title="Cerco"
+              onAdd={() => vm.handleOpenService('cerco_eletronico')}
+            />
+            <ServiceAddCard
+              title="Imagem"
+              onAdd={() => vm.handleOpenService('busca_por_imagem')}
+            />
+            <ServiceAddCard
+              title="Placas correlatas"
+              onAdd={() => vm.handleOpenService('placas_correlatas')}
+            />
+            <ServiceAddCard
+              title="Placas conjuntas"
+              onAdd={() => vm.handleOpenService('placas_conjuntas')}
+            />
+            <ServiceAddCard
+              title="Reserva de imagem"
+              onAdd={() => vm.handleOpenService('reserva_de_imagem')}
+            />
+            <ServiceAddCard
+              title="Análise de imagem"
+              onAdd={() => vm.handleOpenService('analise_de_imagem')}
+            />
+            <ServiceAddCard
+              title="Outros"
+              onAdd={() => vm.handleOpenService('outros')}
+            />
+          </div>
+
+          <div className="mt-4 space-y-4">
+            <ServiceList
+              label="Busca por placa"
+              fields={vm.buscaPorPlaca.fields}
+              onRemove={vm.buscaPorPlaca.remove}
+              onEdit={(idx) =>
+                vm.openServiceModalForEdit('busca_por_placa', idx)
+              }
+              renderRow={() => null}
+            />
+
+            <ServiceList
+              label="Busca por radar"
+              fields={vm.buscaPorRadar.fields}
+              onRemove={vm.buscaPorRadar.remove}
+              onEdit={(idx) =>
+                vm.openServiceModalForEdit('busca_por_radar', idx)
+              }
+              renderRow={() => null}
+            />
+
+            <ServiceList
+              label="Cerco eletrônico"
+              fields={vm.cercoEletronico.fields}
+              onRemove={vm.cercoEletronico.remove}
+              onEdit={(idx) =>
+                vm.openServiceModalForEdit('cerco_eletronico', idx)
+              }
+              renderRow={() => null}
+            />
+
+            <ServiceList
+              label="Busca por imagem"
+              fields={vm.buscaPorImagem.fields}
+              onRemove={vm.buscaPorImagem.remove}
+              onEdit={(idx) =>
+                vm.openServiceModalForEdit('busca_por_imagem', idx)
+              }
+              renderRow={() => null}
+            />
+
+            <ServiceList
+              label="Placas correlatas"
+              fields={vm.placasCorrelatas.fields}
+              onRemove={vm.placasCorrelatas.remove}
+              onEdit={(idx) =>
+                vm.openServiceModalForEdit('placas_correlatas', idx)
+              }
+              renderRow={(idx) => (
+                <CorrelataListForm
+                  register={vm.register}
+                  control={vm.control}
+                  index={idx}
+                  name="placas_correlatas"
+                  disabled={vm.isLoading}
+                />
+              )}
+            />
+
+            <ServiceList
+              label="Placas conjuntas"
+              fields={vm.placasConjuntas.fields}
+              onRemove={vm.placasConjuntas.remove}
+              onEdit={(idx) =>
+                vm.openServiceModalForEdit('placas_conjuntas', idx)
+              }
+              renderRow={(idx) => (
+                <CorrelataListForm
+                  register={vm.register}
+                  control={vm.control}
+                  index={idx}
+                  name="placas_conjuntas"
+                  disabled={vm.isLoading}
+                />
+              )}
+            />
+
+            <ServiceList
+              label="Reserva de imagem"
+              fields={vm.reservaDeImagem.fields}
+              onRemove={vm.reservaDeImagem.remove}
+              onEdit={(idx) =>
+                vm.openServiceModalForEdit('reserva_de_imagem', idx)
+              }
+              renderRow={() => null}
+            />
+
+            <ServiceList
+              label="Análise de imagem"
+              fields={vm.analiseDeImagem.fields}
+              onRemove={vm.analiseDeImagem.remove}
+              onEdit={(idx) =>
+                vm.openServiceModalForEdit('analise_de_imagem', idx)
+              }
+              renderRow={() => null}
+            />
+
+            <ServiceList
+              label="Outros"
+              fields={vm.outros.fields}
+              onRemove={vm.outros.remove}
+              onEdit={(idx) => vm.openServiceModalForEdit('outros', idx)}
+              renderRow={() => null}
+            />
+          </div>
+        </Section>
+
+        <Section
+          title="Locação interna"
+          isOpen={vm.openSections.internal}
+          onToggle={() => vm.toggleSection('internal')}
+        >
+          <div className="space-y-3">
+            <div className="space-y-1.5">
+              <Label className={styles.fieldLabel}>Equipe</Label>
+              <Controller
+                control={vm.control}
+                name="equipe_id"
+                render={({ field }) => (
+                  <Select
+                    value={field.value ?? ''}
+                    onValueChange={(v) => field.onChange(v || null)}
+                    disabled={vm.isLoading}
+                  >
+                    <SelectTrigger className={`h-11 ${styles.inputBg}`}>
+                      <SelectValue placeholder="Selecione" />
+                    </SelectTrigger>
+                    <SelectContent className={styles.selectContentForm}>
+                      {MOCK_TEAMS.map((t) => (
+                        <SelectItem
+                          key={t.id}
+                          value={t.id}
+                          className={styles.selectItemForm}
+                        >
+                          {t.title}
+                        </SelectItem>
+                      ))}
+                    </SelectContent>
+                  </Select>
+                )}
+              />
+              {vm.errors.equipe_id?.message && (
+                <p className="text-xs text-destructive">
+                  {vm.errors.equipe_id.message}
+                </p>
+              )}
+            </div>
+
+            <div className="space-y-2">
+              <Label className={styles.fieldLabel}>Prioridade</Label>
+              <div className="grid grid-cols-3 gap-3">
+                <PriorityButton
+                  active={vm.watch('prioridade') === 'URGENTE'}
+                  label="Urgente"
+                  onClick={() => vm.setValue('prioridade', 'URGENTE')}
+                  disabled={vm.isLoading}
+                />
+                <PriorityButton
+                  active={vm.watch('prioridade') === 'ALTA'}
+                  label="Alta"
+                  onClick={() => vm.setValue('prioridade', 'ALTA')}
+                  disabled={vm.isLoading}
+                />
+                <PriorityButton
+                  active={vm.watch('prioridade') === 'ROTINA'}
+                  label="Rotina"
+                  onClick={() => vm.setValue('prioridade', 'ROTINA')}
+                  disabled={vm.isLoading}
+                />
+              </div>
+            </div>
+          </div>
+        </Section>
+
+        <Section
+          title="Adicionar comentário"
+          isOpen={vm.openSections.comment}
+          onToggle={() => vm.toggleSection('comment')}
+        >
+          <Textarea
+            placeholder="Escreva um comentário"
+            disabled={vm.isLoading}
+            className={`${styles.fakeEditor} ${styles.inputBg}`}
+            {...vm.register('comentario_inicial')}
+          />
+        </Section>
+
+        <Section
+          title="Anexar documentos"
+          isOpen={vm.openSections.attachments}
+          onToggle={() => vm.toggleSection('attachments')}
+        >
+          <div className={styles.attachmentsLayout}>
+            <div className={styles.attachmentsDocumentList}>
+              {vm.files.length === 0 ? (
+                <p className={styles.uploadBoxHint}>Nenhum arquivo anexado.</p>
+              ) : (
+                <div className={styles.fileList}>
+                  {vm.files.map((f, idx) => (
+                    <div key={`${f.name}-${idx}`} className={styles.fileRow}>
+                      <SquareCheck
+                        className={`${styles.fileRowCheckIcon} shrink-0`}
+                        aria-hidden
+                      />
+                      <p className={styles.fileRowFileName} title={f.name}>
+                        {f.name}
+                      </p>
+                      <Button
+                        type="button"
+                        variant="ghost"
+                        className={`${styles.fileRowDeleteBtn} h-8 w-8 shrink-0 p-0`}
+                        onClick={() => vm.removeFile(idx)}
+                        disabled={vm.isLoading}
+                        title="Excluir anexo"
+                      >
+                        <Trash className="h-4 w-4" aria-hidden />
+                      </Button>
+                    </div>
+                  ))}
+                </div>
+              )}
+            </div>
+
+            <div className={styles.uploadColumn}>
+              <label className={styles.uploadBox}>
+                <input
+                  className="hidden"
+                  type="file"
+                  multiple
+                  accept=".pdf,.doc,.docx"
+                  onChange={(e) => vm.onDropFiles(e.target.files)}
+                  disabled={vm.isLoading}
+                />
+                <Upload className="h-6 w-6 shrink-0 text-[var(--tc-icon-subtle)]" />
+                <span className={styles.uploadBoxText}>
+                  Clique para fazer upload ou arraste o arquivo
+                </span>
+                <span className={styles.uploadBoxHint}>
+                  PDF, DOC, DOCX (máx. 10MB)
+                </span>
+              </label>
+            </div>
+          </div>
+        </Section>
+
+        <div className={styles.footerBar}>
+          <div className={styles.footerInner}>
+            <Button
+              type="button"
+              variant="secondary"
+              className={styles.cancelButton}
+              disabled={vm.isLoading}
+              onClick={() => vm.reset(vm.defaultValues)}
+            >
+              Cancelar
+            </Button>
+
+            <Button
+              type="submit"
+              disabled={vm.isLoading}
+              className={styles.saveButton}
+            >
+              {vm.isLoading ? 'Salvando...' : 'Salvar'}
+            </Button>
+          </div>
+        </div>
+      </form>
+
+      <ServiceModal
+        serviceModalOpen={vm.serviceModalOpen}
+        editIndex={vm.serviceModalEditIndex}
+        closeServiceModal={vm.closeServiceModal}
+        initialBuscaPorPlaca={initialBuscaPorPlaca}
+        initialBuscaPorRadar={initialBuscaPorRadar}
+        initialCerco={initialCerco}
+        initialBuscaPorImagem={initialBuscaPorImagem}
+        initialPlacasCorrelatas={initialPlacasCorrelatas}
+        initialPlacasConjuntas={initialPlacasConjuntas}
+        initialReservaImagem={initialReservaImagem}
+        initialAnaliseImagem={initialAnaliseImagem}
+        initialOutros={initialOutros}
+        onSaveBuscaPorPlaca={(value, editIndex) => {
+          const normalized = {
+            plate: value.plate ?? null,
+            period_start: nullIfEmpty(value.period_start),
+            period_end: nullIfEmpty(value.period_end),
+          }
+
+          if (editIndex !== null) {
+            vm.buscaPorPlaca.update(editIndex, normalized)
+          } else {
+            vm.buscaPorPlaca.append(normalized)
+          }
+
+          vm.closeServiceModal()
+        }}
+        onSaveBuscaPorRadar={(value, editIndex) => {
+          const normalized = {
+            plate: value.plate,
+            period_start: nullIfEmpty(value.period_start),
+            period_end: nullIfEmpty(value.period_end),
+            radar_address: nullIfEmpty(value.radar_address),
+          }
+
+          if (editIndex !== null) {
+            vm.buscaPorRadar.update(editIndex, normalized)
+          } else {
+            vm.buscaPorRadar.append(normalized)
+          }
+
+          vm.closeServiceModal()
+        }}
+        onSaveCerco={(value, editIndex) => {
+          const normalized = {
+            plate: value.plate,
+            vehicle_observations: nullIfEmpty(value.vehicle_observations),
+          }
+
+          if (editIndex !== null) {
+            vm.cercoEletronico.update(editIndex, normalized)
+          } else {
+            vm.cercoEletronico.append(normalized)
+          }
+
+          vm.closeServiceModal()
+        }}
+        onSaveBuscaPorImagem={(value, editIndex) => {
+          const normalized = {
+            plate: nullIfEmpty(value.plate),
+            period_start: nullIfEmpty(value.period_start),
+            period_end: nullIfEmpty(value.period_end),
+            address: nullIfEmpty(value.address),
+            description: nullIfEmpty(value.description),
+          }
+
+          if (editIndex !== null) {
+            vm.buscaPorImagem.update(editIndex, normalized)
+          } else {
+            vm.buscaPorImagem.append(normalized)
+          }
+
+          vm.closeServiceModal()
+        }}
+        onSavePlacasCorrelatas={(value, editIndex) => {
+          const normalized = {
+            interest_interval_minutes: value.interest_interval_minutes,
+            detection_count: value.detection_count,
+            detection: value.detection,
+            items: value.items.map((item) => ({
+              plate: item.plate,
+              period_start: nullIfEmpty(item.period_start),
+              period_end: nullIfEmpty(item.period_end),
+            })),
+          }
+
+          if (editIndex !== null) {
+            vm.placasCorrelatas.update(editIndex, normalized)
+          } else {
+            vm.placasCorrelatas.append(normalized)
+          }
+
+          vm.closeServiceModal()
+        }}
+        onSavePlacasConjuntas={(value, editIndex) => {
+          const normalized = {
+            interest_interval_minutes: value.interest_interval_minutes,
+            detection_count: value.detection_count,
+            detection: value.detection,
+            items: value.items.map((item) => ({
+              plate: item.plate,
+              period_start: nullIfEmpty(item.period_start),
+              period_end: nullIfEmpty(item.period_end),
+            })),
+          }
+
+          if (editIndex !== null) {
+            vm.placasConjuntas.update(editIndex, normalized)
+          } else {
+            vm.placasConjuntas.append(normalized)
+          }
+
+          vm.closeServiceModal()
+        }}
+        onSaveReservaImagem={(value, editIndex) => {
+          const normalized = {
+            period_start: nullIfEmpty(value.period_start),
+            period_end: nullIfEmpty(value.period_end),
+            orientation: nullIfEmpty(value.orientation),
+          }
+
+          if (editIndex !== null) {
+            vm.reservaDeImagem.update(editIndex, normalized)
+          } else {
+            vm.reservaDeImagem.append(normalized)
+          }
+
+          vm.closeServiceModal()
+        }}
+        onSaveAnaliseImagem={(value, editIndex) => {
+          const normalized = {
+            period_start: nullIfEmpty(value.period_start),
+            period_end: nullIfEmpty(value.period_end),
+            orientation: nullIfEmpty(value.orientation),
+          }
+
+          if (editIndex !== null) {
+            vm.analiseDeImagem.update(editIndex, normalized)
+          } else {
+            vm.analiseDeImagem.append(normalized)
+          }
+
+          vm.closeServiceModal()
+        }}
+        onSaveOutros={(value, editIndex) => {
+          const normalized = {
+            orientation: nullIfEmpty(value.orientation),
+          }
+
+          if (editIndex !== null) {
+            vm.outros.update(editIndex, normalized)
+          } else {
+            vm.outros.append(normalized)
+          }
+
+          vm.closeServiceModal()
+        }}
+      />
+    </div>
+  )
+}
